refactoring of the server part: change implementation to impl
