+++ server/src/main/java/com/lunchtime/controllers/MenuItemDishController.java
@@ -9,8 +9,7 @@ import java.util.Optional;
 
 import javax.validation.Valid;
 
-import com.lunchtime.implementation.MenuItemDishServiceImplement;
-import com.lunchtime.models.Feedback;
+import com.lunchtime.service.impl.MenuItemDishServiceImplement;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -20,7 +19,6 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
 import com.lunchtime.models.MenuItemDish;
-import com.lunchtime.service.MenuItemDishService;
 
 @RestController
 @RequestMapping("/api/menuitemdish")
