+++ server/src/main/java/com/lunchtime/models/Dish.java
@@ -1,17 +1,11 @@
 package com.lunchtime.models;
 
 import java.util.List;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToMany;
-import javax.persistence.OneToMany;
+import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
+
 
 @Getter
 @Setter
